move compiler helpers to helpers file to avoid circular imports
